requirements.txt
requests
discord.py
termcolor
./smccd_directory_client
./smccd_schedule_client